pyproject.toml
[tool.poetry]
name = "django-tutorial"
version = "0.1.0"
description = "Django Pokemon Tutorial"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"
amqp = "^5.3.1"
annotated-types = "^0.7.0"
asgiref = "^3.8.1"
celery = "^5.4.0"
django-cors-headers = "^4.7.0"
django-environ = "^0.12.0"
django-ipware = "^7.0.1"
django-ninja = "^1.0.1"
django-structlog = "^9.0.1"
flower = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"